pyproject.toml
[tool.poetry]
name = "echobot"
version = "0.1.0"
description = ""
authors = ["Eric Lara <ericzon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
botbuilder-integration-aiohttp = ">=4.15.0"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
botbuilder-core = "^4.16.2"
botframework-connector = "^4.16.2"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"
openai = "^1.59.9"
langgraph = "^0.2.67"
langchain-openai = "^0.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
